Move concatenate_vectors into its own header
The common library is now used by all targets in sycl-blas
