+++ cmake/CmakeFunctionHelper.cmake
@@ -131,6 +131,7 @@ endforeach(executor)
 add_library(${func} OBJECT ${FUNC_SRC})
 set_target_compile_def(${func})
 target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_link_libraries(${func} PRIVATE sycl_blas_common)
 message(STATUS "Adding SYCL to target ${func}")
 add_sycl_to_target(TARGET ${func} SOURCES ${FUNC_SRC})
 endfunction(generate_blas_unary_objects)
@@ -179,6 +180,7 @@ endforeach(executor)
 add_library(${func} OBJECT ${FUNC_SRC})
 set_target_compile_def(${func})
 target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_link_libraries(${func} PRIVATE sycl_blas_common)
 message(STATUS "Adding SYCL to target ${func}")
 add_sycl_to_target(TARGET ${func} SOURCES ${FUNC_SRC})
 endfunction(generate_blas_binary_objects)
@@ -230,6 +232,7 @@ endforeach(executor)
 add_library(${func} OBJECT ${FUNC_SRC})
 set_target_compile_def(${func})
 target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_link_libraries(${func} PRIVATE sycl_blas_common)
 message(STATUS "Adding SYCL to target ${func}")
 add_sycl_to_target(TARGET ${func} SOURCES ${FUNC_SRC})
 endfunction(generate_blas_binary_special_objects)
@@ -282,7 +285,8 @@ foreach(executor ${executor_list})
 endforeach(executor)
 add_library(${func} OBJECT ${FUNC_SRC})
 set_target_compile_def(${func})
-target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${SYCLBLAS_COMMON_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_link_libraries(${func} PRIVATE sycl_blas_common)
 message(STATUS "Adding SYCL to target ${func}")
 add_sycl_to_target(TARGET ${func} SOURCES ${FUNC_SRC})
 endfunction(generate_blas_ternary_objects)
@@ -625,6 +629,7 @@ add_library(${func} OBJECT ${gemm_sources})
 set_target_compile_def(${func})
 # The blas library depends on FindComputeCpp
 target_include_directories(${func} PRIVATE ${SYCLBLAS_SRC} ${SYCLBLAS_INCLUDE} ${THIRD_PARTIES_INCLUDE})
+target_link_libraries(${func} PRIVATE sycl_blas_common)
 message(STATUS "Adding SYCL to target ${func}")
 add_sycl_to_target(TARGET ${func} SOURCES ${gemm_sources})
 endfunction(generate_blas_gemm_objects)
@@ -684,6 +689,7 @@ function(generate_quantize)
     ${SYCLBLAS_INCLUDE}
     ${ComputeCpp_INCLUDE_DIRS}
     ${COMPUTECPP_SDK_INCLUDE})
+  target_link_libraries(quantize PRIVATE sycl_blas_common)
   message(STATUS "Adding SYCL to target quantize")
   add_sycl_to_target(TARGET quantize SOURCES ${FUNC_SRC})
 endfunction(generate_quantize)
