+++ common/include/common/common_utils.hpp
@@ -17,16 +17,11 @@
 #include <vector>
 
 #include "benchmark_cli_args.hpp"
-//#include "blas_meta.h"
-//
-//#include "utils/quantization.hpp"
 
-#include <common/clwrap.h>
-
-#ifdef BLAS_VERIFY_BENCHMARK
+#include <common/vector_utils.hpp>
+#include <common/clwrap.hpp>
 #include <common/float_comparison.hpp>
 #include <common/system_reference_blas.hpp>
-#endif
 
 using index_t = int;
 using blas1_param_t = index_t;
